[feat]: auth 기능 정리
