+++ loaders/expressLoader.js
@@ -1,7 +1,7 @@
 import cookieParser from 'cookie-parser'
 import bodyParser from 'body-parser'
 import morgan from 'morgan'
-import api from '../api'
+import api from '../api/routes'
 
 const expressLoader = (app) => {
   app.use(cookieParser())
